requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
pillow==11.1.0
Pygments==2.19.1
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.1.0
qrcode==8.1
rich==13.9.4
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
